Rewrite as remote plugin to avoid blocking
